+++ docs/source/user_guide/changelog.rst
@@ -28,6 +28,7 @@ Unreleased
 ==========
 
 - Remove parameter ``IDENTITY_TRANSFER`` setting it to ``True`` if a knowledge transfer folder is provided.
+- Fix bug in the :ref:`validator_reference` when saving session from the pop up dialog at closing the GUI.
 
 6.0.12
 ======